ocr-api-v2/requirements.txt
# Framework Web
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart

# PaddleOCR
paddlepaddle
paddleocr

# Traitement d'images
#opencv-python-headless
Pillow
PyMuPDF
numpy

# Utilitaires
aiofiles
structlog